// apps/super-app/src/libs/sanity/sanity.types.ts
export interface AiToolFaqItemsByLocale {
  _type: "aiToolFaqItemsByLocale";
  en?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  zh?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  th?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  ar?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  es?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  ko?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  ja?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
  hi?: ({
    _key: string;
  } & AiToolFaqQaPair)[];
}

export interface AiToolFaqQaPair {
  _type: "aiToolFaqQaPair";
  question?: string;
  answer?: {
    children?: {
      marks?: string[];
      text?: string;
      _type: "span";
      _key: string;
    }[];
    style?: "normal" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "blockquote";
    listItem?: "bullet" | "number";
    markDefs?: {
      href?: string;
      _type: "link";
      _key: string;
    }[];
    level?: number;
    _type: "block";
    _key: string;
  }[];
  answerPlain?: string;
}
// ...
export interface AiToolFAQs {
  _id: string;
  _type: "aiToolFAQs";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  itemsByLocale?: AiToolFaqItemsByLocale;
}